Refactor interactive deployment scripts into modules
