+++ scripts/tools/interactive-deploy/test-import.js
@@ -1,0 +1,2 @@
+import inquirer from 'inquirer';
+console.log('Inquirer imported successfully:', !!inquirer);
